Remove 'group-active-tab' component and refine routing
