+++ projects/dev/src/app/group/group-content/group-content.component.html
@@ -50,4 +50,3 @@
 <ng-template #noPermissionBlock>
   <p>The user doesn't have enought permission to see this content</p>
 </ng-template>
-<app-group-active-tab (activeTabChange)="onActiveTabChange($event)"></app-group-active-tab>
