+++ projects/dev/src/app/group/group-content/group-content.component.ts
@@ -1,12 +1,19 @@
-import { Component, OnInit, ElementRef, EventEmitter, Input, Output } from '@angular/core';
+import { Component, OnInit, ElementRef, Input } from '@angular/core';
 import * as _ from 'lodash';
-import { GroupService } from '../../shared/services/api/group.service';
 import { ActivatedRoute, Router, UrlSegment } from '@angular/router';
-import { Group, initializeGroup } from '../../shared/models/group.model';
+import { Group } from '../../shared/models/group.model';
+import { Location } from '@angular/common';
 
 export enum Management {
   None = 'none',
   MembershipsAndGroup = 'memberships_and_group'
+}
+
+export enum TabUrls {
+  Overview = 'overview',
+  Composition = 'members',
+  Administration = 'managers',
+  Settings = 'settings'
 }
 
 @Component({
@@ -16,24 +23,33 @@ export enum Management {
 })
 export class GroupContentComponent implements OnInit {
 
-  group: Group;
-  groupId = 0;
+  @Input() group: Group;
   activeTab = 0;
 
   constructor(
     private activatedRoute: ActivatedRoute,
     private elementRef: ElementRef,
-    private groupService: GroupService,
-    private router: Router) {
-    this.group = initializeGroup();
+    private router: Router,
+    private location: Location) {
   }
 
   ngOnInit() {
-    this.activatedRoute.params.subscribe((routeParams) => {
-      this.groupId = routeParams.id;
-      this.groupService.getGroup(this.groupId).subscribe((group) => {
-        this.group = group;
-      });
+    this.activatedRoute.url.subscribe(_ => {
+      const path = this.location.path().split('/').pop();
+      switch (path) {
+        case TabUrls.Composition:
+          this.activeTab = 1;
+          break;
+        case TabUrls.Administration:
+          this.activeTab = 2;
+          break;
+        case TabUrls.Settings:
+          this.activeTab = 3;
+          break;
+        default:
+          this.activeTab = 0;
+          break;
+      }
     });
   }
 
@@ -60,16 +76,16 @@ export class GroupContentComponent implements OnInit {
 
     switch (iTab) {
       case 0:
-        this.router.navigate([`/dev/groups/${this.groupId}`]);
+        this.router.navigate([`/dev/groups/${this.group.id}`]);
         break;
       case 1:
-        this.router.navigate([`/dev/groups/${this.groupId}/members`]);
+        this.router.navigate([`/dev/groups/${this.group.id}/members`]);
         break;
       case 2:
-        this.router.navigate([`/dev/groups/${this.groupId}/managers`]);
+        this.router.navigate([`/dev/groups/${this.group.id}/managers`]);
         break;
       case 3:
-        this.router.navigate([`/dev/groups/${this.groupId}/settings`]);
+        this.router.navigate([`/dev/groups/${this.group.id}/settings`]);
         break;
     }
   }
