+++ projects/dev/src/app/group/group-manage/group-manage.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { StatusService } from '../../shared/services/status.service';
 import { GroupService } from '../../shared/services/api/group.service';
-import { Group } from '../../shared/models/group.model';
+import { Group, initializeGroup } from '../../shared/models/group.model';
 import { Member } from '../../shared/models/member.model';
 
 @Component({
@@ -11,7 +11,7 @@ import { Member } from '../../shared/models/member.model';
   styleUrls: ['./group-manage.component.scss'],
 })
 export class GroupManageComponent implements OnInit {
-  groupdata = {};
+  group: Group;
   groupId;
   status;
 
@@ -53,21 +53,16 @@ export class GroupManageComponent implements OnInit {
     private activatedRoute: ActivatedRoute,
     private statusService: StatusService,
     private groupService: GroupService
-  ) {}
+  ) {
+    this.group = initializeGroup();
+  }
 
   ngOnInit() {
     this.activatedRoute.params.subscribe((routeParams) => {
       this.groupId = routeParams.id;
 
       this.groupService.getGroup(this.groupId).subscribe((group: Group) => {
-        this.groupdata = {
-          ID: group.id,
-          name: group.name,
-          type: group.type,
-          grades: [group.grade],
-          date: group.created_at,
-          description: group.description
-        };
+        this.group = group;
       });
     });
